read this commit
OK I corrected the bug in the optimized version of the integrator. Now it can use Eigen::Ref to automatically extract the K part from the Xi vector and we do not make use anymore of hat in the SO(3) integrator. 
However, for the other integrator, we still need to compute hat_K at all the (first layer) quadrature points. However this could be optimized in some ways if I get to find a way to avoid hat_K*hat_K. Then we won.